third_party/blink/web_tests/fast/css/webkit-mask-crash-implicit-expected.txt

PASS document.styleSheets[1].rules[0].style.cssText is "-webkit-mask-repeat-x: repeat; -webkit-mask-repeat-y: inherit;"
PASS document.styleSheets[1].rules[0].style.getPropertyValue('-webkit-mask') is ""
PASS document.styleSheets[1].rules[0].style.getPropertyValue('-webkit-mask-repeat') is ""
PASS successfullyParsed is true

TEST COMPLETE
